requirements.txt
fastapi==0.124.2
uvicorn[standard]==0.38.0
python-multipart==0.0.20
python-dotenv==1.2.1